Fixed Intake Tilt and SP/RX Imports
There was something wrong with the intake tilt class.
Also you need to fix the switch with "case null" in the elevator class. I don't think you can do that. I didn't touch it. There are about 5 instances of this error.
You also didn't have the SP/RX's imported in the Elevator class, fixed that as well.